add timeout for connection and some s/die/_exit
Crashing is some situation is not a good idea.  If the other process
is dead, it has probably already crashed, so crashing again would
overwrite the core file, and in every case it doesn't tell us much.
It's more interesting to see why the first process crashed in the
first place.