switch to phos_uri
Initially I took the IRI parser from gmid, stripped of the UTF-8
things and transformed back into a URI parser.  Then I've added some
functions to resolve URIs, but it wasn't enough.  gmid parser is aimed
at servers, and so it doesn't handle very well relative links.

This makes use of the libphos uri parser, which was coded to follow
the RFC more strictly, and to fully support all kinds of URIs.  It'll
need to be expanded in the future and be improved for sure