generate a new id for every request
bad things (tm) can happen if we do a IMSG_GET hwen a page is
loading.  This fixes it by generating a new id for every request, so
the current one (if there is one) can safely finish (stop_tab can take
a while) while we're doing the next one.