retire restore_cursor and the force-redraw optimization (for now)
don't consider restore_cursor as a public interface and while there
rename it as restore_curs_x, as it describes better the behaviour of
the function.

I have to disable the buffer->force-redraw optimization, at least from
now, given that cmd.c don't update curs_y anymore, and with that
optimisation in place the cursor y position doesn't get updated upon
redraw.

Add a comment as a future note though.