fix two related bug
Don't crash when there's a certificate mismatch and after saving the
cert, before choosing to persist it, the page gets redirected; also,
don't repeat the "Certificate mismatch" after the first yes, if
there's a redirect.  Now when the user accept a certificate, is
treated as temporary saved for the whole session.