optimization: don't endwin() when we can simply move the windows
After a resize we should do the endwin+refresh+clear dance, but in
other situations, like when entering the minibuffer, we can simply
re-wrap the page and move the windows.  This prevents a small flash
when entering the minibuffer, as well as fewer redraw (redraw after
the command and shortly after because it was scheduled).