don't crash on invalid tab ids in handle_imsg_buf
My original idea was to crash when receiving data for an invalid tab.
This idea just doesn't hold well in an asynchronous context.  If we
change page before the current tab finished loading we issue tab_stop,
but even after that we still could receive a IMSG_BUF for that old tab
id.