refactor net code
Generalize a bit the flow.  Expect a request to be written and a body
to be sended with net_send_ui.  This will make easier to add other
protocols like finger or gopher in the future.  Also, switch to
evbuffer instead of manually calling tls_read/write: with evbuffers we
can abstract over the actual I/O routine and just read/write buffers
in memory.  This too will help with adding more protocols.