handle IMSG_GET_FILE: serve local files
IMSG_GET_FILE handles a path to a local file.  fs now behaves like
net, it sends a IMSG_GOT_CODE/META and optionally the file.

This also includes a bit of refactoring: the send_file function was
extracted from handle_get so it can be used in handle_get_file too.