proposal: show custom message when the capsule is not reachable
As discussed in the mailing list, this is a quick draft for a (imho)
neat functionality: if there's an error connecting to a capsule, try
to do a TXT query at _offline.$hostname (this is subject to change)
and show the message to the user.

It's a nice way to communicate to the user that a certain capsule is,
for example, supposed to be offline in certain situations.