switch from ^L to a line of only `-'
The ^L implementation was ugly and wrong.  Using a character from the
already allowed set is better.

Also, the motivation for ^L was to avoid breaking existing pages (as
much as possible).  But using a line consisted of only `-' shouldn't
break anything, but only improve the legibility.  In the future other
characters may be added (like a line of only EN DASH or EM DASH for
example.)