dirname fix
dirname(3) is really one of the worst part of POSIX, portability-wise.
Implementors can decide to either modify the passed string or return a
pointer to an internal storage.

The current code fails when dirname returns a pointer to an internal
storage because between the strcmp call and the mkdir we go into a
recursion that can modify `path'.  We can't use copy either because
*some* implementation of dirname would have changed that.

How to escape from this hell?  Take yet another copy of the path!