add landlock support on linux
landlock is applied only to the ui process to drop fs access and in the
fs process to limit where telescope can read/write files.

The network process is more difficult to landlock because while in
theory it doesn't need *any* fs access, in practice it needs to read (at
least) files inside /etc/ for DNS to work.