provide HOSTCC and HOSTCFLAGS variables to simplify cross-compilation
pagebundler is a helper that needs to be built with the *host* compiler
because it's used to generate some sources.  In normal situations,
HOSTCC and HOSTCFLAGS are just ${CC} and ${CFLAGS}, but folks that
cross-compile can specify there the host compiler and flags to
successfully build telescope.

The idea is largely based on what gforth does: it re-uses ${CFLAGS} for
the host compiler though, while I'm adding an extra variable for that.