use download_cols to wrap the text in the download buffer
download_lines is a very small value, for a normally sized terminal is
exactly 5.  This was the cause behind the download pane glitch, 5 was
used as *column number* for the reflow.

Now, to be honest, the exact width passed to wrap_page is not
important.  wrap_page will only wrap the size string, which we know is
less than or equal to FMT_SCALED_STRSIZE-1 (6).  We could also
hardcode the value eventually, but using download_cols reads better.