move parse_last_session to the fs process
~/.telescope/session was parsed in the main process before dropping in
the sandbox: that's no good.  This moves the initialization of the ui
events later in ui_main_loop and makes so that the fs process, after
entering the sandbox, parses and sends the content of the last session
back to the main process; when it's done the ui gets loaded and
telescope is back rocking.

I'm plannig to re-use the struct session_tab during the save session
operation soon.