allow to re-open closed tabs
Save closed tabs to a "killed tabs" queue, for up to a max of
"max-killed-tabs", and allow to re-open them.  Keep the tab history
but not its content.

This integrates nicely with the session file, just add the "killed"
flag as alias for TAB_KILLED and put it in the ktabshead queue upon
initalization.  It's thus possible to re-open closed tabs even after
telescope has been closed, and tabs keeps their history!

Add a new cmd, cmd_tab_undo_close, to try to re-open the most
recently closed tab and bind it to `u'.

While here, add a TODO for the tab history memory leak.