provide a way to bypass the cache
This changes the last argument of load_url to be a `mode' bitmap
instead of a boolean ``nohist''.  LU_MODE_NOHIST is the old 1,
LU_MODE_NONE is provided just for readability and the new LU_MODE_NOCACHE
allows to bypass the caching when loading a page.

Telescope only uses the cache when:

 - re-opens a closed tab
 - navigates backward and forward in the history

All other situations result in the usual network request.