save and restore scroll position
Save the offset to the first and current line of the buffers and restore
it when the page is loaded.  This restores almost the exact viewport: we
save the offset of the *real* line in the document, but they get wrapped
before being displayed, so the actual restored position may be slightly
above from the original one (think of a long paragraph that gets wrapped
over multiple visual lines.)  In practice, this isn't a drawback but
only an improvement over the current behaviour.

In the future we could save also the offset *inside* the line, to
restore the exact position, but that felt unnecessary for now.

Persisting the postion on the session file ensures that even when
re-opening telescope the scroll position is correct.

There are some very active pages, such as antenna or other aggregator,
where this feature doesn't really work well if you re-open the page
after a while, since new links are available, but otherwise is a god
sent!

See also github issue #1