parsers: add serialize function
Add a parser_serialize function and an optional function `serialize' for
parsers.  The default implementation just prints the text line, which is
suitable for parser_textplain and parser_textpatch; gemtext and
gophermap uses their own specific unparse functions.

This is intended for an incoming cache refactoring: we'll cache the page
serialized in the correct format, not once it has been translated to
gemtext.  This will allow in the future to provide commands such as
``save page to disk'' or interface with an on-disk cache.