set_scroll_position: always set current_line
There's a small chance we may end up finding the topline but not the
current line if a page content is abruptedly changed.

While here also drop `curfound = 1': it's a dead assignment since it
returns in that conditions.

Bug found while reading the output of clang' scan-build, but not thanks
to clang.  scan-build complained about the `curfound' variable, and by
re-reading the code I spotted the missing case.