don't highlight the first completion by deafult
This is the first commit of a series to improve the handling of
completions.  Currently it's a mess: some commands look at the
selected entry, others at what was typed in the minibuffer... it's not
clear which commands does what.  So, change of defaults: don't
highlight anything by default to avoid confusing the user, C-n and C-p
now are the obvious way to move the focus from the minibuffer to a
completion.

This doesn't change how the commands handle the selection, that will
done in a follow-up commit.