draft of the new iri parser
for the moment is just a URI (RFC3986) parser, and not even a complete
one, but I like it much more than phos_uri.  It's more gentle in the
memory consumption and has a cleaner API.

It's not complete, some parts are still stubbed or completely to
implement, but it's the start.  It compiles.

Step-by-step integration of it in telescope and testing will come later.