iri: rewrite remove_dot_segments to be clearer
the previous version was straight out RFC3986 (more or less, it had a
bug I didn't spot), but it's ugly and extremely long.  Rewrite it to be
somewhat inspired by canonpath (from kern_pledge.c) and way simpler.
There are some small differences in how UNIX paths and URLs paths are
canonized (in UNIX path you don't care too much of a trailing '/', in
URLs you do.)

The new code is more clear and passes all the tests in RFC3986, so I'm
calling it a win.