add load-url-use-heuristic knob
This is still WIP and may change before the next release.

It could be worthwile to resolve the URLs in load-url using the
current page as base, so that while browsing <gemini://a.org/b/c>,
using load-url to go to "/" would load <gemini://a.org/> instead of
<file:///>.  The new option `load-url-use-heuristic', if unset,
provides exactly this behaviour.

It is still a bit rough to use, as for e.g. loading <d.net> would now
yield <gemini://a.org/b/d.net> instead of <gemini://d.net>.

Will need more work in-tree, but for now add it so we can reason about
it.

See github issue #10