apply load-url-use-heuristic for command line arguments too
This semplifies a bit the handling.  humanify_url() now takes the base
URL which we can then use when not using the heuristics.

Command-line arguments now take an implicit base of <file://$PWD/>
when the heuristics are disabled, so that foo.gmi resolves to the
local file even without <./>.

See github issue #10.