remove phos; now unused
phos was an attempt at building a gemini client/server library.  It
didn't went as I expected, mostly because it's such a simple
protocol... Anyway, the only use in telescope was for the uri module,
which has already been replaced with the built-in iri.  Which still
sucks (and despite the name, can still only handle URIs) but it's
hopefully better.  The APIs are better.