add boolean variables
The 'if zero/nonzero then...' is a bit annoying.  Add proper booleans
to the grammar of the config and use them when appropriate.  Makes the
description and the review of configuration files easier.

This means that the old

	set olivetti-mode = 1

can now be written as

	set olivetti-mode = true

Compatibility with the previous integer values will be kept for a
while.