bp_select: call exit_minibuffer() when done
Currently it is still fine to access ministate after
exit_minibuffer().  In the future, if we ever support recursive
minibuffers it won't.

So, defer the exit_minibuffer() call and use minibuffer_compl_text()
instead of peeking into ministate.

While here, also add the "Bookmarked" message when it successfully
bookmarked the page.