draft of telescope-identity(1)
This is intended to generate and manage certificates outside of
telescope.  Some operations are already available in the telescope
UI, such as use and forget a certificate for a specific page, but
generate and removal are hard at the moment to implement inside
telescope limited UI, so provide an helper program for that.