change default patch/diff colours to magenta and cyan
red/green is not a great choice for diff colours: it's a bad combo
for (some) colour-blind people, and they may also carry an implicit
meaning of 'good' and 'bad', or 'ok' and 'not ok'.  Instead, just
like tog and gotwebd, use a combo of purple and cyan, which was
choosen by a colour-blind user (I'm told).