libgrapheme: don't allow parallel builds
When building the bundled version of libgrapheme, don't build it in
parallel, with "-j" is used to "make".

In most cases, there's no guarantee that libgrapheme is going to be
built before telescope, and the detection for this is such that it's
always going to be called on-demand, which almost always guarantees that
libgrapheme will be built after telescope.