tls: guard against faulty gemini servers
Some gemini servers have faulty TLS handling, such as:

  gg: tls_close: EOF without close notify

In such cases, telescope will receive that error and naturally assume
there was a problem and stop processing.  However, in the case of there
still being legitimate content read from the request, this should still
be displayed, rather than ignored.

This change therefore detects this condition and adds a status of 'W' to
telescope's modeline to indicate that the information presented is from
a faulty gemini server and could be truncated and hence display oddly.