ui: add capability to run external cmds
When telescope encounters a file type which it cannot display it prompts
the user whether they wish to save the file, or not.

This patch extends this idea -- if the file is saved, then the user is
prompted to run a command against that downloaded file.

This controllable via the 'external-cmd' setting.  It defaults to
'xdg-open'.

This is based off work originally started by op@.