point out that yacc is a dependency
It's just easier to always require yacc, which is a fairly small
build-dependency, rather then the current wording.  Also, this
allows us to have the configure check which is more useful and
friendlier than the previous behaviour of sometimes failing midway
thru the build due to a missing yacc.

See https://codeberg.org/telescope-browser/telescope/pulls/7