sandbox: temporarly disable unveil(2) on OpenBSD
I'd like to use execlp() in the near future, which means that unveil
needs to be disabled.  It's already virtually disabled, since we
can shell out to execute stuff, so it's not a great loss.

This will be improved in the future once the "core" process will
be split off the UI.