add mini-edit-external
it's bound by default to C-x in the minibuffer keymap.  It'll open
an external editor to edit the content of the minibuffer.  When the
editor exits, we'll call the completion function with the content
of the file.

Stuff inputted with an external editor are not persisted to the
history thought.  That would require the UI to handle multiline
strings which is currently impossible.