mailcap: add parser for external command handling
This introduces a (limited) subset of mailcap handling, as defined in
RFC 1524.

This makes it possible for telescope to run commands on different mime
types as defined in a mailcap file, but defaults to a default mime type
of using xdg-open if no match is found, or no mailcap file defined.