configure: add a --with-default-editor option
Ths is useful for Linux distributions which might not solely rely on
EDITOR/VISUAL being set in the environment, but would rather use a
wrapper command.

Debian prefers sensible-editor(1), for example.

Rather than make hard-coded assumptions in telescope for each
distribution, this will make it easier for packagers of telescope to
set this if needed.

Defaults to ed(1) if not set.