code: add wrappers for common failure cases
Certain calls to well-known functions sometimes result in similar
error handling across the code base.  Sometimes there's a need to
centralise the failure case.

Achieve this for common wrappers to often-used functions.

The transformation is achieved via coccinelle, commited with this patch.

Note that coccinelle can be told to ignore code blocks with comment
delimiters.

This transformation applies just to telescope's code, and not any of the
bundled depdencies in telescope (libgrapheme).