mailcap: improve flag handling
Extra information present on a mailcap line can extend to almost
anything.  It cannot be assumed that flags such as "copiousoutput" or
"needsterminal" will be the first set of flags after the command field.

For now, treat everything after the "command" part of a mailcap entry as
potential fields, and still scan for "copiousoutput" and
"needsterminal".