minibuffer: redesign enter_minibuffer and fix regression
use a struct to pass the arguments to enter_minibuffer since they're
so many.  While here, also allow to pass a printf(3)-like format string
for the prompt and specify the initial input.

Also, fix a regression: if we don't set vline.cplen we can't properly
handle the input, so do that in enter_minibuffer to avoid having to fix
all the callers.