handle tab characters
tab characters have their width depending on the column they're in,
since they extend to the next multiple of 8.  (citation needed?)

So, keep track of the column when considering the length (in columns)
of the text, so that we can render them properly.

In the future we might want to turn them into spaces (either at read
or render time) just to stay on the safe side in case not all
terminals/ncurses implementations use 8 columns.