build: macos: set PKG_PATH_CONFIG programatically
Under certain conditions, set PKG_PATH_CONFIG programatically, as in,
find all the relevant occurances under /opt/homebrew

This is necessary because some CI systems might not do this, and it's a
pain having to enumerate dependencies of dependencies.

There is a slight downside that ./configure might pause for a few
seconds.