parser_serialize: fflush()
If the amount of data being written to the FILE pointer ends up being a
lot, it's worth fflush()ing on it here, rather then relying on specific
callers to do so.  This can happen if telescope is fetching HTML
content from an HTTP proxy, for example.

This helps to improve writes to disk, especially via the write-buffer
command.